menuTagListのenumを分離
