+++ lib/model/menu_tag.dart
@@ -1,0 +1,13 @@
+enum MenuTag {
+  japanese("和"),
+  western("洋"),
+  chinese("中華"),
+  ethnic("エスニック"),
+  cool("さっぱり"),
+  oily("こってり"),
+  luxury("豪華"),
+  quick("時短");
+
+  final String name;
+  const MenuTag(this.name);
+}
