--- lib/model/tag.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Tag {
-  String title;
-  Color backgroundColor;
-
-  Tag({
-    required this.title,
-    required this.backgroundColor,
-  });
-}
